Initial commit
Starting by adding the build machinery, an empty main file for the jupiter tool,
a skeleton test and... an initial implementation of decimal7.

decimal7 is a small roll-my-own floating point decimal implementation. Created
it myself to avoid use of some complex drop-in replacement from an external
library since the STL doesn't seem to have one yet. Also it means I can
customise it as I please.

It's probably not the best implementation, but it suits my needs (being able to
parse prices from betfair without running into HORRIBLE floating point issues).

And so it begins.