decimal7: Normalise input so 1.230 == 1.23
This resolves an important issue - floating point numbers with trailing zeroes
after the decimal place are equivalent. It is redundant to store these and worse
it breaks equality tests. We expect the following:

decimal7(1230, 2) == decimal7(123, 1)

Failing to do so might result in some rather hard to debug issues.

This patch additionally separates out the normalisation code into its own
function for neatness.