decimal7: Fix sign/unsigned interaction in .fract()
The .fract() method is retrieving the raw integral value which can be negative.
In order to avoid implicitly converting a negative signed int64 to an unsigned
int64 and the HORROR or what C++ does with that, we have to explicitly check for
the sign and absolute the value before modding.