decimal7: Update the .mult10*() implementation
Convert the existing method to .mult10n() with an exception thrown in case of
invalid value requested rather than a broken and incorrectly described mask.

Add some helper functions for req_places = 2 and 3 as these are common use
cases.

Again, tests will be added as part of a larger test change.