sajson: Use int64 everywhere for integer values
Despite the JSON standard technically not supporting int64, it is very useful
to be able to read integers as integers even when they exceed INT_MAX (or are
less than INT_MIN) without first casting to a double wherein you endure the
horrors of floating point.

This patch attempts to provide the minimum changeset to implement this, simply
upgrading TYPE_INTEGER to represent int64.

The get_int53_value() function is now misnamed, but this will be addressed in a
subsequent patch.

Some of the original unit tests required changing to reflect the fact we can now
correctly handle int64's.