sajson: Rename get_int53_value() to reflect fact it can return int64
If something has caused the number to be stored as a double (e.g. decimal place
or exponent) then we revert to the int53 logic as this is the maximum integer
a double can represent.

However if it is an actual integer, we can handle this directly. Update the name
of the function and add a comment to make this clear.