sajson: Add decimal support
This patch adds a new decimal type to sajson using the janus::decimal7 type.

The implementation parses numbers with a decimal place as decimal by default,
however it extends .get_double_value() to accept a decimal so existing users of
sajson should still be able to use the code as before.

It additionally extends .try_get_integer_value() to handle decimals as well.

The .get_decimal_value() can accept integers as well as decimals, which is very
convenient when you want a number in decimal format no matter whether the data
happens to be encoded as an integer or decimal.

In the cases where a decimal cannot represent the number (with 1 specific
exeption) the value will be stored as a double. This occurs f the input value
exceeds decimal::MAX or is less than decimal::MIN, or uses an exponent. The case
where a decimal is used when the number cannot be represented correctly is if
the decimal place count exceeds the maximum representable. At this point
trailing decimal places are stripped by the decimal implementation.

In future revisions the implementation could be updated to handle smaller
exponents as decimal (or even an integer if appropriate), convert oversized
fractions to double and add a mode where we never use double (to avoid odd
edgecases in input data).

Original unit tests have been updated as necessary (but minimally) to reflect
the type change, specific decimal unit tests will be added in a subsequent
commit.