dynamic_buffer: Add dynamic buffer implementation
This implements a simple allocator which consists of a pre-allocated chunk of
memory which we can append to. It does not permit dynamic allocation (perhaps
ironic given the name), but is dynamic in that it is expandable and can be reset
to reclaim all memory used.

The purpose is to be able to have a binary buffer that can be accessed directly
but to be able to allocate memory very simply without slow system-level
allocations.

Everything is aligned to 64-bit values (i.e. 8 bytes) to avoid any weird
performance killing unaligned accesses.

Unit tests will be added in a subsequent commit.