dynamic_buffer: Add .add_raw() method to add arbitrary memory ranges to buffer
The add_raw() method adds the ability to copy from any block of memory at a byte
granularity which makes the dynamic buffer far more flexible and acts as the
basis for a number of future intended uses (e.g. constructing a binary file
format).

The method copies the data into the buffer, filling any trailing unaligned bytes
with zeroes.