dynamic_buffer: Add .read_offset() getter and .read_uint64() function
This fleshes out the read offset concept with an actual usage and exposes the
read offset to users.

It additionally adds a check_read_overflow() private method to ensure that we
don't read off the end of the buffer and genericises the throw logic to avoid
code duplication.

Finally, it adds a .reset_read() method to reset the read offset to the start of
the buffer.