parse: Add function to parse ISO-8601 timestamps
ISO-8601 timestamps are of the format '2020-03-11T13:20:00.123Z' or without
milliseconds '2020-03-11T13:20:00Z'. The 'Z' refers to Zulu timezone (GMT)
though this can be set to different values.

We choose to only parse Zulu, and explicitly check for the trailing Z to make
sure the value is in this timezone.

We try to be relatively fast in parsing the timestamp, using lookup tables to do
so.

The function has some very simple tests but generally expects, if the string is
of the correct length and has a training 'Z' that the values will be valid.

This is to avoid too much time spent checking values. This might be something I
revisit on a future revision to this code.

See https://en.wikipedia.org/wiki/ISO_8601 for more details on ISO-8601.