dynamic_buffer: Handle zero-size strings
Previously zero-sized strings would cause an additional uint64 to bbe allocated
due to an additional byte of storage being requested to store a null terminator.

This patch specifically checks for empty strings on read/add string.

This is actually useful in practice if a string parameter being written to the
buffer is empty and we want to express the empty string correctly.