style: Disable buggy clang-tidy unused function test
This check seems to be bugged out as it repeatedly claims internal::is_leap() is
unused when in fact it is.

I am not enabling cppcheck's (which doesn't seem buggy) as at this stage in the
project there are many unused functions. I will enable this at a later stage.