json: Add betfair metadata header parser
Retrieves the pertinent _static_ contents of the betfair metadata into the
format of a header struct. I will probably expand this later to include other
fields, but starting with the fundamentals.

This patch also adds a test which focuses on the code populating the header as
expected rather than the internal workings of the JSON parser, value parser,
etc. which are already tested elsewhere.