dynamic_buffer: Fix bug with .read_string() not reading null terminator
Previously, the .read_string() method would simply read the count of characters
which excludes the null terminator character. Since we are aligning to uint64
words, this can cause a big issue where the read offset is  off-by-1.

Fix the issue and update the unit test to specifically check for this.