json: Add unit test for betfair::parse_meta_json()
We already added betfair::parse_meta_json() bit-by-bit as we were adding
internal functions which this function is essentially calling sequentially.

This adds a specific unit test which ensures it does what is expected. This test
doesn't have to be overly extensive as the internal function tests have already
gone deep on their implementation.