neptune: Add initial implementation using betfair metadata JSON parser
Add a simple implementation that simply reads a number of specified metadata
JSON files and parses them into binary file format before writing the total
collection of data to the specified output file.

This is helpful in being able to test against the dataset, once upgraded to also
read data and to be a little more sophisticated I will add end-to-end tests.