price_range: Add a class which provides betfair price ranges
'Price' here means decimal odds. These have variable tick sizes depending on
price, so we need to be able to navigate through a valid range and match, both
fuzzily and precisely, a price to an index in the price range.

It is convenient to, where possible, refer to indexes within the array of valid
betfair prices rather than the price itself, so this class provides the ability
to map from one to the other.

Additionally, we prefer to operate with integers, and as the maximum decimal
places of any price at betfair is 2, we simply multiply prices by 100 to obtain
an integral value. The price range class defaults to using this representation,
though it allows for conversion to floating point values too when required.