price_range: Add methods to convert floating point price to index/pricex100
Add methods .price_to_nearest_index() and .price_to_nearest_pricex100() to add
the ability to convert a floating point price to a specific index for the
nearest price.

Typically I try to avoid using floating points wherever possible so in the
ordinary operation of janus this would not be needed, however if performing
calculations that explicitly work with floating points, e.g. looking at implied
odds, then it can be useful to be able to convert back and forth using these.