ladder: Add initial runner price ladder class
This adds a new ladder class which represents data that applies to a specific
runner's price ladder. The initial implementation stores only unmatched volume
as well as tracking minimum ATL and maximum ATB and total unmatched ATL and ATB
values.

Later revisions of this will include additional ladder-specific data such as
matched volume.

ATB is Available-To-Back and means an existing, unmatched, back bet, therefore
requiring a LAY to match it. Equivalently, ATL is Available-To-Lay and means an
existing, unmatched, lay bet, requiring a BACK to match it.

We store backs and lays on the same ladder with backs stored as positive volumes
and lays as negative volumes.

The ladder methods (and unit tests!) explicitly check to ensure discontinuities
do not occur, e.g. a back bet appearing in the middle of lay bets or vice-versa.

However, replacement of a minimum ATL or a maximum ATB is allowed in order to
avoid requiring a clear to be sent before a set when the spread moves.