update: Generate MARKET_CLEAR update when "img":true
When "img":true is found in the betfair stream data this means the update is a
snapshot and therefore we need to clear down existing market data and set it to
what is contained in the update.

In order to signify this we generate a MARKET_CLEAR message as the first update after a possible MARKET_ID update.