universe: Only apply market, runner timestamp updates when they are updated
The initial implementation simply set the LAST market/runner timestamp to the
one we just received, but this is incorrect as we may well be updating a
separate market/runner.

We intentionally do NOT set timestamp when ADDING a market/runner, as in fact
adding should be immediately followed with data, and if no data is present then
actually no update has really occurred.