update: Strip duplicate IDs
When processing data I have saved many of the updates are purely to bdatl/bdatb
alone therefore given we are now only processing atb/atl these end up as empty
updates. We then often get a number of update messages that are just runner IDs,
one after the other.

These are meaningless as only the last runner ID update applies.

This patch checks for this specifically and strip any unneeded runner IDs as we
parse JSON.