config: Add config helpers to read configurable fields
In order to use the betfair API we need to know username, password, unattended
login certificate paths and application key. This is something that should be
kept in a separate file so we need to be able to parse these things from a
configuration. Given we already have the capability to read JSON effectively it
makes sense to use JSON to achieve this.