betfair: Add intitial betfair connection object
This object brings together the logic I have been working on in order to allow
network operations against the betfair API. It uses the [non-interactive][0]
login using user-provided certificates which can be specified by the
configuration file.

It currently only implements login and logout but forms the basis for adding
further functionality.

[0]:https://docs.developer.betfair.com/display/1smk3cen4v3lu3yomq5qye0ni/Non-Interactive+%28bot%29+login