betfair: stream: Add market subscription and read line methods
The market subscription method uses string allocations, etc. but again as we
only expect to call this very infrequently it is not an issue.

The read line implementation however is performance-critical - here we use an
internal buffer where we can cache lines of input given that we can read
multiple lines at once.

At this point it's possible to implement streaming (roughly :)