betfair: session: Add ability to get data until disconnect
Default betfair API calls to this mode. This is because of course betfair decide
they will send you a large chunk of data without any Content-Length to troll
you.

So to work around this, add the ability to wait until we are simply disconnected
(thankfully they do this by default).

Really we should be checking for Content-Length and if not present using this
mode, but it seems a convention that the API disconnects you after the query so
this is OK for now.