betfair: api: Retrieve metadata when getting market IDs
The same API is used for both finding market IDs and getting the metadata. Also
we need to split our metadata queries into blocks according to [0], for the data
we request this is a weighting of 2 up to a max of 200, so our maximum query count is 100.

This results in a multi-line metadata response each containing separate arrays
of metadata, we can update neptune to handle this correctly.

[0]:https://docs.developer.betfair.com/display/1smk3cen4v3lu3yomq5qye0ni/Market+Data+Request+Limits