jupiter: Add initial non-skeleton implementation
This is a very rough-and-ready implementation but forms a good basis for further
work and a working data recorder so we can resume recording data using
non-legacy code.

It blocks on writing I/O to disk, is totally synchronous and probably has a
dozen other issues I've not even thought about but it's a start! :)