neptune: Implement modern metadata retrieval
Implement intended new approach for metadata retrieval using the new format
metadata files being retrieved by jupiter.

This is still incomplete as are not yet processing stream data but it does
load/save state into a neptune.db file so we have the basis for the stream data
implementation in place.