neptune: Place in loop with signal handler for ctrl+C
Add a simple loop with a sleep for the time being, later we'll add inotify
support. This allows this to run as a daemon constantly updating binary files.

This is also a MVP to provide consistently updated binary data, later jupiter
will transmit this data as a stream in a pubsub architecture.