bet: Handle unmatched volume correctly on runner removal
If a bet has an adjustment factor applied due to a runner removal, then the
unmatched portion needs to be dealt with according to betfair rules.

Unmatched lays are just cancelled, unmatched backs are left intact.

Since we are scaling the matched portion of the bet, if there is any unmatched
portion then we need to split that out.

Rather than make things too complicated, just update the .apply_adj_factor()
method to return true in the case that the unmatched portion needs to be split
out and set a reference out param indicating the stake size of the split out
bet.