db: Load all market updates into a string before a dynamic buffer
Since we're doing I/O this copying is not going to add all that much more
overhead and this opens the door to reading market updates into a correctly
sized container.

Since this is directly invoked by the other read_market_updates() overload
there's no need to add a specific test - if it were incorrect so would the other
overload be.