neptune: Sort updates chronologically
It seems there was a significant bug previously where, because the DB is stored
as an unordered map, and we retrieved stream json filenames from it in arbitrary
order, we'd end up putting market updates in incorrect order as a result.

Sort the ID list in order to get correct ordering.