analyse: Add initial implementation
This a tool which allows backtesting on multicore systems against a filtered
subset of market binary data.

This opens the door to serious analysis and is the key under-the-bonnet system
for apollo.

The initial implementation is pretty messy and has horrible levels of
indentation etc. along with introducing log output on test (ugh), but it is
intended to be a MVP.